Add interface group split horizon support
This optionally restricts network policy processing to only pod-facing
interfaces. Other traffic passing through the node will not be affected
by it even if it matches network policies. This is useful if your nodes
also act as routers for cluster traffic not destined for them.

Needs a patched CNI or other mechanisms to set the interface group on
the relevant links.