Fix checking of egress rules after ingress accepts
Traffic should only be allowed if both ingress and egress rules allow
it. Before this fix, a packet which was accepted by ingress rules was
immediately accepted without checking the egress rules. But when both
the source and destination pod are on the same node, the same packet may
need to be checked against both ingress and egress rules.

The simplest way to fix this is to use separate base chains for ingress
and egress. This means that, when an ingress rule accepts, the egress
chain is still evaluated.

This fixes many cyclonus test cases, for example "ingress: allow all".