Fix named port lookup
The named port lookup was completely broken. The register offsets were
correct for IPv6 but not IPv4. Instead of dynamically adjusting the
offset of the protocol and port, I moved the IP address to the end which
is simpler to implement.

Additionally, the endianness handling for protocol and port was wrong.
I'm still not sure if the endianness handling for the protocol is
correct now. It works on little endian hosts, but I suspect that it
might not work on big endian. But I don't have a big endian host to test
it on, so I'm leaving it as is.

This fixes most of the cyclonus tests with tag named-port.