Fix bug in namespace update
There was a logic bug in reevalPodInRule, which meant that if a pod
selector matches but the pod was already selected previously, the
reference was removed instead of leaving it as is.

This bug can only be observed with a policy which has multiple namespace
selectors, where one of the selectors matches both before and after the
namespace update, but the other selector only matches before or after.
For example, with the following policy, adding or removing the `demo:
demo` label from the default namespace demonstrates the bug:

```
apiVersion: networking.k8s.io/v1
kind: NetworkPolicy
metadata:
  name: multi-ns-selector
spec:
  podSelector: {}
  ingress:
  - from:
    - namespaceSelector:
        matchLabels:
          kubernetes.io/metadata.name: default
    - namespaceSelector:
        matchLabels:
          demo: demo
  policyTypes:
  - Ingress
```

There is no existing cyclonus test which catches this bug, but I have
written one which does. I still need to figure out where to put this
test, maybe we can get it upstreamed to cyclonus.