Fix selector evaluation when namespace not loaded
There was a bug in the evaluation of pod selectors: If the namespace
selector is Nothing, but the namespace is not yet loaded, then the
selector was treated as not matching any pods. But when that namespace
is later loaded, the selector is not reevaluated because the namespace
selector is Nothing. During initial loading, namespaces, pods and rules
can be loaded in any order, and this order should not affect how traffic
is filtered.

The fix is to only check whether the namespace exists if the namespace
selector is not Nothing. In this case, the selector will be reevaluated
once the namespace is loaded.