Handle very long Kubernetes names
nftables chain and set names are limited to 256 characters. If there are
namespaces, pods or policies with very long names, this can break the
policy controller. To fix this, use the UID instead when the combined
namespace and name is longer than 128 bytes. We still want to use the
name when possible, because that is more debugging friendly.