Fix missing elements in IPv6 sets
Sets which had the same keytype on both v4 and v6 were empty on v6. The
reason for this is that, in AddSet, when s.KeyType6 is not set, it will
use s.KeyType for v6 instead. But this is only set to s.v6.KeyType,
while s.KeyType6 remains unset, meaning that s.KeyType6.Bytes is 0.
Instead, splitVals needs to use s.v6.KeyType.

This fixes cyclonus tests with port ranges on IPv6.